Add generic health platform service and create healthkit and healthconnect service that implements it
